First stab at py2/3 compatibility

+++ blingalytics/sources/__init__.py
@@ -49,13 +49,15 @@ all filters accept the following optional standard arguments:
   :doc:`/widgets` for more.
 """
 
+from past.builtins import basestring
+from builtins import object
 from datetime import timedelta
 from decimal import Decimal
 
 from blingalytics import formats
 
 
-ADD_TYPES = (int, long, Decimal, float, timedelta)
+ADD_TYPES = (int, int, Decimal, float, timedelta)
 
 class Source(object):
     """
